refactor: 상금별 로또 개수 계산 로직 변경
Map이 아닌 List를 사용하도록 변경하였습니다. 그리고 상금이 3억이 넘는것을 고려하여 long타입으로 변경하였습니다.

+++ src/main/java/lotto/configuration/Prize.java
@@ -12,7 +12,7 @@ public enum Prize {
 
     final private int matchCount;
     final private Boolean matchBonus;
-    final private int prizeMoney;
+    final private long prizeMoney;
 
     Prize(int matchCount, Boolean matchBonus, int prizeMoney) {
         this.matchCount = matchCount;
@@ -31,7 +31,7 @@ public enum Prize {
         return matchCount;
     }
 
-    public int getPrizeMoney() {
+    public long getPrizeMoney() {
         return prizeMoney;
     }
